Move boltDB mirgrator to the separate package
