+++ cmd/di.go
@@ -46,6 +46,7 @@ import (
 	"github.com/mysteriumnetwork/node/core/state"
 	"github.com/mysteriumnetwork/node/core/storage/boltdb"
 	"github.com/mysteriumnetwork/node/core/storage/boltdb/migrations/history"
+	"github.com/mysteriumnetwork/node/core/storage/boltdb/migrator"
 	"github.com/mysteriumnetwork/node/eventbus"
 	"github.com/mysteriumnetwork/node/feedback"
 	"github.com/mysteriumnetwork/node/firewall"
@@ -388,7 +389,7 @@ func (di *Dependencies) bootstrapStorage(path string) error {
 		return err
 	}
 
-	migrator := boltdb.NewMigrator(localStorage)
+	migrator := migrator.NewMigrator(localStorage)
 	err = migrator.RunMigrations(history.Sequence)
 	if err != nil {
 		return err
